==> 2 - Recopilar artistas ganadores.ipynb <==
import openpyxl

def leer_excel(nombre_archivo, artistas, contador):
    # Abre el archivo Excel
    libro = openpyxl.load_workbook(nombre_archivo)

    # Selecciona la primera hoja del libro (puedes cambiar el índice si hay múltiples hojas)
    hoja = libro.active

    # Recorre las filas y columnas
    for fila in hoja.iter_rows():
        for celda in fila:
            if celda.value is not None and 'Year' not in celda.value and 'Best' not in celda.value:
                artistas.add(celda.value)
                if celda.value in contador:
                    contador[celda.value] += 1
                else:
                    contador[celda.value] = 1

    # Cierra el archivo Excel
    libro.close()

# Leemos los archivos Excel de cada edición para sacar una lista común de artistas y premios ganados por cada uno
archivos = ['62nd Annual GRAMMY Awards 2019.xlsx',
            '63rd Annual GRAMMY Awards 2020.xlsx',
            '64th Annual GRAMMY Awards 2021.xlsx',
            '65th Annual GRAMMY Awards 2022.xlsx',
            '66th Annual GRAMMY Awards 2023.xlsx']

contador_premios = {} # Contador de premios por artista; cada vez que se repita un artista significa que ha ganado un premio más
artistas_unicos = set() # Conjunto de ganadores; para saber el total de ganadores
for nombre_archivo in archivos:
    print(f'Archivo: {nombre_archivo}')
    leer_excel('ficheros/'+nombre_archivo, artistas_unicos, contador_premios)
    print(f'- Artistas descargados: {len(artistas_unicos)}\n-> {artistas_unicos}')
    print(f'-> Contador premios: {contador_premios}\n')

# Guardamos en un nuevo archivo Excel todos los ganadores de las ediciones de los años 2019-2023
wb = openpyxl.Workbook()
hoja_activa = wb.active
hoja_activa.title = "Ganadores 19-23"

# Iteramos sobre la lista de premios y escribe los datos en columnas
for indice, dato in enumerate(contador_premios.items(), start=1):
    ganador, premios = dato
    hoja_activa.cell(row=indice, column=1, value=ganador) # Escribe el ganador en la primera columna
    hoja_activa.cell(row=indice, column=2, value=premios) # Escribe el número de premios en la segunda columna

# Guardamos el libro de Excel en función del año que estemos estudiando
wb.save('Ganadores_19-23.xlsx')
